immediate-sync : Monitor directories for create events
This change will be able to create watcher objects for the directories
to notify the IN_CREATE along with IN_CLOSE_WRITE events and sync to the
sibling BMC.

IN_CREATE will catch when directories created in the watching
directory, resulting in creating watchers for the created sub
directories too as inotify is not supporting the recursive monitoring
of a directory.

Tested:

Verified the following in the simics model.

- The directories configured from the JSON was monitoring for
  IN_CREATE and IN_CLOSE_WRITE events. The creation or modifications
  of files inside the watching directories are handled by IN_CREATE
  and IN_CLOSE_WRITE events.

- The creation of subdirectories inside a watching directory is
  notified using IN_CREATE and watcher is added to the created sub
  directory too.

Change-Id: I93e791327a69e115f0ea28b975bf6163cb76bbd1
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>