inotify : process the IN_MOVED_TO signal
This commit introduce the changes to process the IN_MOVED_TO inotify
events if emits from the filesystem.

In rysnc layer, when the directory get syncs to the sibling BMC as
result of IN_MOVED_TO, the action is to copy the data which is
moved to the directory. Hence the DataOp enum value is DataOps::COPY

Example: When a PEL is deleted via a DBus method, the corresponding
PEL files are moved from
`/var/lib/phosphor-logging/extensions/pels/logs/` to
`/var/lib/phosphor-logging/extensions/pels/logs/archive/`. This
operation triggers an IN_MOVED_TO event in the `archive` directory,
which is now processed and used to sync the data to the sibling BMC.

Change-Id: I7e7debc3c6ff870904e6a434609927bb8f98372c
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>