Implement the ExcludeList option in dataWatcher class
The commit implements the logic to process and implement the
ExcludeList configured in the data-sync config.

    - If the `ExcludeList` option has directories, skip those
      directories and their child's also from adding inotify watches.
    - If the `ExcludeList` option has files, skip the inotify events
      received for them from processing.

In data-sync, the ExcludeList will help users to configure the list of
data paths inside a configured directory which don't want to monitor
for data changes, whereas majority of the data need to monitor.

Additional Changes : Fixed the wrong KEY in the dataWatcher class log.

Tested:
Tested and confirmed the following scenarios in huygens simics model:

- Configured single and multiple directories in ExcludeList
    - No inotify watches get added for the directories in ExcludeList.
    - Receiving and processing the inotify events for all other files
      and directories.
    - While syncing, Rsync CLI is properly framing the exclude list
      using "--filter='-/ <ABSOLUTE_PATH>'"

- Configured single and multiple files in ExcludeList
    - Data sync is skipping the inotify events for the configured
      files from processing.

Change-Id: If4285af543ed774c72e5713726f4f5aca4b5dc8b
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>