Implement the IncludeList option in data-sync layer
The commit implements the logic to process and implement the
IncludeList configured in the data-sync config and the implementation
does the following :

    - If IncludeList is configured for a configured path, while
      syncing the path(during full sync and periodic sync) rsync will
      use only the paths configured in the includeList as source paths.
    - During immediate sync, data-sync will be able to know the exact
      modified path and hence the modified path will be used as source
      path by rsync, instead of using all paths in includeList.
    - If IncludeList has directories as configured , the data-sync
      will add watches only for those directories and their child
      directories.
    - If the IncludeList has files, process only the inotify events
      received for those files and skip others.

In data-sync, the IncludeList will help the users to configure path
names in IncludeList if they want to monitor only few number of data
paths inside a directory and rest of the files don't need to monitor.

Note : Although rsync supports the `--filter` option to selectively
include paths, it requires explicitly including all parent directories
as well, and excluding the rest. This makes the CLI command
more complex. Hence, the above approach has been chosen.

Tested:
Tested and confirmed the following scenarios in huygens simics model:

- Configured single and multiple directories in IncludeList
    - Inotify watches get added only for those directories in the
      IncludeList.
- Configured single and multiple files in ExcludeList
    - Inotify watches get added only for those files in the
      IncludeList.

Change-Id: I7332aadc309e2f6a15aaa77cd2940d80293b3590
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>