inotify : Process IN_MOVED_FROM signal
This commit introduce the changes to process the IN_MOVED_FROM inotify
events if emits from the filesystem.

In rysnc layer, when the directory get sync to the sibling BMC as
result of IN_MOVED_FROM , the action is to remove the data which is
moved from. Hence the DataOp enum value is DataOps::DELETE

Example: When a PEL is deleted via a DBus method, the corresponding
PEL files are moved from
`/var/lib/phosphor-logging/extensions/pels/logs/` to
`/var/lib/phosphor-logging/extensions/pels/logs/archive/`.  This move
operation triggers an `IN_MOVED_FROM` event in the `logs` directory,
which is now captured and used to sync the data to the sibling BMC.

Change-Id: Ia0e0c7dda259f764615cbd2d30acd2bbdfc61246
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>