rsync-cli : Add `--relative` to preserve folder structure
This commit adds support for the --relative flag in rsync. With this
flag enabled, the full path of each specified file or directory is
preserved and sent to the server, rather than just the final path
component. Hence in rsync command, the destination path need to
configure explicitly only if it is different from the source path.

In phosphor-data-sync the `--relative` flag addresses the following
scenarios:

 - Automatic creation of intermediate directories:
    When the full source path includes multiple nested directories,
    --relative ensures that missing parent or grandparent directories
    are automatically created at the destination.

    Example: If the source path is /a/b/c/d/e/ and the destination only
    contains /a/b/, syncing directory e with --relative will result in
    automatic creation of c and d under /a/b/.

 - Restoring data from backup directories:
    For data copying into backup directories, preserving the full source
    path helps to identify the original destination during data
    restoration from backup dir after a failover.

Additionally, this commit includes necessary changes in unit tests to
accommodate changes introduced by enabling the --relative option.

Change-Id: I14e9062f01929d43369e11c545c00bc25a0bb1d0
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>