rsync-cli : Add --relative flag
This commit adds support for the --relative flag in rsync. With this
flag enabled, the full path of each specified file or directory is
preserved and sent to the server, rather than just the final path
component.

Additionally, this commit includes necessary changes in unit tests to
accommodate changes introduced by enabling the --relative option.

Change-Id: I14e9062f01929d43369e11c545c00bc25a0bb1d0
Signed-off-by: Arya K Padman <aryakpadman@gmail.com>